Reuse getColumns
