+++ packages/admin-panel/src/VizBuilderApp/components/PreviewSection.js
@@ -14,6 +14,7 @@ import { useReportPreview } from '../api';
 import { usePreviewData, useVisualisation, useVizConfig, useVizConfigError } from '../context';
 import { JsonEditor } from './JsonEditor';
 import { IdleMessage } from './IdleMessage';
+import { getColumns } from '../../utilities';
 
 const PreviewTabs = styled(MuiTabs)`
   background: white;
@@ -105,34 +106,6 @@ const TABS = {
 };
 
 const getTab = index => Object.values(TABS).find(tab => tab.index === index);
-
-const convertValueToPrimitive = val => {
-  if (val === null) return val;
-  switch (typeof val) {
-    case 'object':
-      return JSON.stringify(val);
-    case 'function':
-      return '[Function]';
-    default:
-      return val;
-  }
-};
-
-const getColumns = ({ columns: columnKeys = [] }) => {
-  const indexColumn = {
-    Header: '#',
-    id: 'index',
-    accessor: (_row, i) => i + 1,
-  };
-  const columns = columnKeys.map(columnKey => {
-    return {
-      Header: columnKey,
-      accessor: row => convertValueToPrimitive(row[columnKey]),
-    };
-  });
-
-  return [indexColumn, ...columns];
-};
 
 export const PreviewSection = () => {
   const [tab, setTab] = useState(0);
