+++ packages/admin-panel/src/utilities/index.js
@@ -12,3 +12,4 @@ export * from './pretty';
 export * from './useDebounce';
 export { checkVisibilityCriteriaAreMet } from './visibilityCriteria';
 export { labelToId } from './labelToId';
+export { getColumns } from './getColumns';
